compute: move `CollectionState` into instance.rs
This commit moves the `CollectionState` type, as well is the
introspection management types it contains into the `instance` module.
They fit much better there, given that the `Instance` controller manages
collections and the `top-level` controller does nothing more than
handing them out to clients. When we move the `Instance` controller into
a separate task, we will make `CollectionState` private so the top-level
controller can't even access it anymore.
