+++ src/compute-client/src/controller/instance.rs
@@ -38,6 +38,7 @@ use mz_storage_types::read_holds::{ReadHold, ReadHoldError};
 use mz_storage_types::read_policy::ReadPolicy;
 use serde::Serialize;
 use thiserror::Error;
+use timely::progress::frontier::{AntichainRef, MutableAntichain};
 use timely::progress::{Antichain, ChangeBatch, Timestamp};
 use timely::PartialOrder;
 use tokio::sync::mpsc;
@@ -48,8 +49,8 @@ use crate::controller::error::{
 };
 use crate::controller::replica::{ReplicaClient, ReplicaConfig};
 use crate::controller::{
-    CollectionState, ComputeControllerResponse, ComputeControllerTimestamp, IntrospectionUpdates,
-    ReplicaId, WallclockLagFn,
+    ComputeControllerResponse, ComputeControllerTimestamp, IntrospectionUpdates, ReplicaId,
+    WallclockLagFn,
 };
 use crate::logging::LogVariant;
 use crate::metrics::{InstanceMetrics, ReplicaCollectionMetrics, ReplicaMetrics, UIntGauge};
@@ -1985,6 +1986,376 @@ where
         self.schedule_collections();
         self.cleanup_collections();
         self.refresh_state_metrics();
+    }
+}
+
+/// State maintained about individual compute collections.
+///
+/// A compute collection is either an index, or a storage sink, or a subscribe, exported by a
+/// compute dataflow.
+#[derive(Debug)]
+pub struct CollectionState<T: Timestamp> {
+    /// Whether this collection is a log collection.
+    ///
+    /// Log collections are special in that they are only maintained by a subset of all replicas.
+    log_collection: bool,
+    /// Whether this collection has been dropped by a controller client.
+    ///
+    /// The controller is allowed to remove the `CollectionState` for a collection only when
+    /// `dropped == true`. Otherwise, clients might still expect to be able to query information
+    /// about this collection.
+    dropped: bool,
+    /// Whether this collection has been scheduled, i.e., the controller has sent a `Schedule`
+    /// command for it.
+    scheduled: bool,
+
+    /// Accumulation of read capabilities for the collection.
+    ///
+    /// This accumulation contains the capabilities held by all [`ReadHold`]s given out for the
+    /// collection, including `implied_read_hold` and `warmup_read_hold`.
+    ///
+    /// NOTE: This field may only be modified by [`Instance::apply_read_hold_changes`] and
+    /// [`Instance::acquire_read_hold`]. Nobody else should modify read capabilities directly.
+    /// Instead, collection users should manage read holds through [`ReadHold`] objects acquired
+    /// through [`Instance::acquire_read_hold`].
+    ///
+    /// TODO(teskje): Restructure the code to enforce the above in the type system.
+    read_capabilities: MutableAntichain<T>,
+    /// A read hold maintaining the implicit capability of the collection.
+    ///
+    /// This capability is kept to ensure that the collection remains readable according to its
+    /// `read_policy`. It also ensures that read holds on the collection's dependencies are kept at
+    /// some time not greater than the collection's `write_frontier`, guaranteeing that the
+    /// collection's next outputs can always be computed without skipping times.
+    implied_read_hold: ReadHold<T>,
+    /// A read hold held to enable dataflow warmup.
+    ///
+    /// Dataflow warmup is an optimization that allows dataflows to immediately start hydrating
+    /// even when their next output time (as implied by the `write_frontier`) is in the future.
+    /// By installing a read capability derived from the write frontiers of the collection's
+    /// inputs, we ensure that the as-of of new dataflows installed for the collection is at a time
+    /// that is immediately available, so hydration can begin immediately too.
+    warmup_read_hold: ReadHold<T>,
+    /// The policy to use to downgrade `self.implied_read_hold`.
+    ///
+    /// If `None`, the collection is a write-only collection (i.e. a sink). For write-only
+    /// collections, the `implied_read_hold` is only required for maintaining read holds on the
+    /// inputs, so we can immediately downgrade it to the `write_frontier`.
+    read_policy: Option<ReadPolicy<T>>,
+
+    /// Storage identifiers on which this collection depends, and read holds this collection
+    /// requires on them.
+    storage_dependencies: BTreeMap<GlobalId, ReadHold<T>>,
+    /// Compute identifiers on which this collection depends, and read holds this collection
+    /// requires on them.
+    compute_dependencies: BTreeMap<GlobalId, ReadHold<T>>,
+
+    /// The write frontier of this collection.
+    write_frontier: Antichain<T>,
+
+    /// Introspection state associated with this collection.
+    collection_introspection: CollectionIntrospection<T>,
+}
+
+impl<T: Timestamp> CollectionState<T> {
+    /// Reports the current read capability.
+    pub fn read_capability(&self) -> &Antichain<T> {
+        self.implied_read_hold.since()
+    }
+
+    /// Reports the current read frontier.
+    pub fn read_frontier(&self) -> AntichainRef<T> {
+        self.read_capabilities.frontier()
+    }
+
+    /// Reports the current write frontier.
+    pub fn write_frontier(&self) -> AntichainRef<T> {
+        self.write_frontier.borrow()
+    }
+
+    fn storage_dependency_ids(&self) -> impl Iterator<Item = GlobalId> + '_ {
+        self.storage_dependencies.keys().copied()
+    }
+
+    fn compute_dependency_ids(&self) -> impl Iterator<Item = GlobalId> + '_ {
+        self.compute_dependencies.keys().copied()
+    }
+
+    /// Reports the IDs of the dependencies of this collection.
+    fn dependency_ids(&self) -> impl Iterator<Item = GlobalId> + '_ {
+        self.compute_dependency_ids()
+            .chain(self.storage_dependency_ids())
+    }
+}
+
+impl<T: ComputeControllerTimestamp> CollectionState<T> {
+    /// Creates a new collection state, with an initial read policy valid from `since`.
+    pub(crate) fn new(
+        collection_id: GlobalId,
+        as_of: Antichain<T>,
+        storage_dependencies: BTreeMap<GlobalId, ReadHold<T>>,
+        compute_dependencies: BTreeMap<GlobalId, ReadHold<T>>,
+        read_holds_tx: mpsc::UnboundedSender<(GlobalId, ChangeBatch<T>)>,
+        introspection_tx: crossbeam_channel::Sender<IntrospectionUpdates>,
+        initial_as_of: Option<Antichain<T>>,
+        refresh_schedule: Option<RefreshSchedule>,
+    ) -> Self {
+        // A collection is not readable before the `as_of`.
+        let since = as_of.clone();
+        // A collection won't produce updates for times before the `as_of`.
+        let upper = as_of;
+
+        // Initialize all read capabilities to the `since`.
+        let implied_read_hold = ReadHold::new(collection_id, since.clone(), read_holds_tx.clone());
+        let warmup_read_hold = ReadHold::new(collection_id, since.clone(), read_holds_tx);
+
+        let mut read_capabilities = MutableAntichain::new();
+        read_capabilities.update_iter(
+            implied_read_hold
+                .since()
+                .iter()
+                .map(|time| (time.clone(), 1)),
+        );
+        read_capabilities.update_iter(
+            warmup_read_hold
+                .since()
+                .iter()
+                .map(|time| (time.clone(), 1)),
+        );
+
+        Self {
+            log_collection: false,
+            dropped: false,
+            scheduled: false,
+            read_capabilities,
+            implied_read_hold,
+            warmup_read_hold,
+            read_policy: Some(ReadPolicy::ValidFrom(since)),
+            storage_dependencies,
+            compute_dependencies,
+            write_frontier: upper.clone(),
+            collection_introspection: CollectionIntrospection::new(
+                collection_id,
+                introspection_tx,
+                initial_as_of,
+                refresh_schedule,
+                upper,
+            ),
+        }
+    }
+
+    /// Creates a new collection state for a log collection.
+    pub(crate) fn new_log_collection(
+        id: GlobalId,
+        read_holds_tx: mpsc::UnboundedSender<(GlobalId, ChangeBatch<T>)>,
+        introspection_tx: crossbeam_channel::Sender<IntrospectionUpdates>,
+    ) -> Self {
+        let since = Antichain::from_elem(timely::progress::Timestamp::minimum());
+        let mut state = Self::new(
+            id,
+            since,
+            Default::default(),
+            Default::default(),
+            read_holds_tx,
+            introspection_tx,
+            None,
+            None,
+        );
+        state.log_collection = true;
+        // Log collections are created and scheduled implicitly as part of replica initialization.
+        state.scheduled = true;
+        state
+    }
+}
+
+/// Manages certain introspection relations associated with a collection. Upon creation, it adds
+/// rows to introspection relations. When dropped, it retracts its managed rows.
+///
+/// TODO: `ComputeDependencies` could be moved under this.
+#[derive(Debug)]
+struct CollectionIntrospection<T> {
+    /// The ID of the compute collection.
+    collection_id: GlobalId,
+    /// A channel through which introspection updates are delivered.
+    introspection_tx: crossbeam_channel::Sender<IntrospectionUpdates>,
+    /// Introspection state for `mz_materialized_view_refreshes`.
+    /// `Some` if it is a REFRESH MV.
+    refresh_introspection_state: Option<RefreshIntrospectionState<T>>,
+}
+
+impl<T> CollectionIntrospection<T> {
+    fn send(&self, introspection_type: IntrospectionType, updates: Vec<(Row, Diff)>) {
+        let result = self.introspection_tx.send((introspection_type, updates));
+
+        if result.is_err() {
+            // The global controller holds on to the `introspection_rx`. So when we get here that
+            // probably means that the controller was dropped and the process is shutting down, in
+            // which case we don't care about introspection updates anymore.
+            tracing::info!(
+                ?introspection_type,
+                "discarding introspection update because the receiver disconnected"
+            );
+        }
+    }
+}
+
+impl<T: ComputeControllerTimestamp> CollectionIntrospection<T> {
+    fn new(
+        collection_id: GlobalId,
+        introspection_tx: crossbeam_channel::Sender<IntrospectionUpdates>,
+        initial_as_of: Option<Antichain<T>>,
+        refresh_schedule: Option<RefreshSchedule>,
+        upper: Antichain<T>,
+    ) -> CollectionIntrospection<T> {
+        let refresh_introspection_state = match (refresh_schedule, initial_as_of) {
+            (Some(refresh_schedule), Some(initial_as_of)) => Some(RefreshIntrospectionState::new(
+                refresh_schedule,
+                initial_as_of,
+                upper,
+            )),
+            (Some(_refresh_schedule), None) => {
+                soft_panic_or_log!("if we have a refresh_schedule, then it's an MV, so we should also have an initial_as_of");
+                None
+            }
+            _ => None,
+        };
+        let self_ = CollectionIntrospection {
+            collection_id,
+            introspection_tx,
+            refresh_introspection_state,
+        };
+
+        if let Some(refresh_introspection_state) = self_.refresh_introspection_state.as_ref() {
+            let insertion = refresh_introspection_state.row_for_collection(collection_id);
+            self_.send(
+                IntrospectionType::ComputeMaterializedViewRefreshes,
+                vec![(insertion, 1)],
+            );
+        }
+
+        self_
+    }
+
+    /// Should be called whenever the write frontier of the collection advances. It updates the
+    /// state that should be recorded in introspection relations, and sends it through the
+    /// `introspection_tx` channel.
+    fn frontier_update(&mut self, write_frontier: &Antichain<T>) {
+        if let Some(refresh_introspection_state) = &mut self.refresh_introspection_state {
+            // Old row to retract based on old state.
+            let retraction = refresh_introspection_state.row_for_collection(self.collection_id);
+            // Update state.
+            refresh_introspection_state.frontier_update(write_frontier);
+            // New row to insert based on new state.
+            let insertion = refresh_introspection_state.row_for_collection(self.collection_id);
+            // Send changes.
+            self.send(
+                IntrospectionType::ComputeMaterializedViewRefreshes,
+                vec![(retraction, -1), (insertion, 1)],
+            );
+        }
+    }
+}
+
+impl<T> Drop for CollectionIntrospection<T> {
+    fn drop(&mut self) {
+        self.refresh_introspection_state
+            .as_ref()
+            .map(|refresh_introspection_state| {
+                let retraction = refresh_introspection_state.row_for_collection(self.collection_id);
+                self.send(
+                    IntrospectionType::ComputeMaterializedViewRefreshes,
+                    vec![(retraction, -1)],
+                );
+            });
+    }
+}
+
+/// Information needed to compute introspection updates for a REFRESH materialized view when the
+/// write frontier advances.
+#[derive(Debug)]
+struct RefreshIntrospectionState<T> {
+    // Immutable properties of the MV
+    refresh_schedule: RefreshSchedule,
+    initial_as_of: Antichain<T>,
+    // Refresh state
+    next_refresh: Datum<'static>,           // Null or an MzTimestamp
+    last_completed_refresh: Datum<'static>, // Null or an MzTimestamp
+}
+
+impl<T> RefreshIntrospectionState<T> {
+    /// Return a `Row` reflecting the current refresh introspection state.
+    fn row_for_collection(&self, collection_id: GlobalId) -> Row {
+        Row::pack_slice(&[
+            Datum::String(&collection_id.to_string()),
+            self.last_completed_refresh,
+            self.next_refresh,
+        ])
+    }
+}
+
+impl<T: ComputeControllerTimestamp> RefreshIntrospectionState<T> {
+    /// Construct a new [`RefreshIntrospectionState`], and apply an initial `frontier_update()` at
+    /// the `upper`.
+    fn new(
+        refresh_schedule: RefreshSchedule,
+        initial_as_of: Antichain<T>,
+        upper: Antichain<T>,
+    ) -> RefreshIntrospectionState<T> {
+        let mut self_ = RefreshIntrospectionState {
+            refresh_schedule: refresh_schedule.clone(),
+            initial_as_of: initial_as_of.clone(),
+            next_refresh: Datum::Null,
+            last_completed_refresh: Datum::Null,
+        };
+        self_.frontier_update(&upper);
+        self_
+    }
+
+    /// Should be called whenever the write frontier of the collection advances. It updates the
+    /// state that should be recorded in introspection relations, but doesn't send the updates yet.
+    fn frontier_update(&mut self, write_frontier: &Antichain<T>) {
+        if write_frontier.is_empty() {
+            self.last_completed_refresh =
+                if let Some(last_refresh) = self.refresh_schedule.last_refresh() {
+                    last_refresh.into()
+                } else {
+                    // If there is no last refresh, then we have a `REFRESH EVERY`, in which case
+                    // the saturating roundup puts a refresh at the maximum possible timestamp.
+                    T::maximum().into()
+                };
+            self.next_refresh = Datum::Null;
+        } else {
+            if timely::PartialOrder::less_equal(write_frontier, &self.initial_as_of) {
+                // We are before the first refresh.
+                self.last_completed_refresh = Datum::Null;
+                let initial_as_of = self.initial_as_of.as_option().expect(
+                    "initial_as_of can't be [], because then there would be no refreshes at all",
+                );
+                let first_refresh = initial_as_of
+                    .round_up(&self.refresh_schedule)
+                    .expect("sequencing makes sure that REFRESH MVs always have a first refresh");
+                soft_assert_or_log!(
+                    first_refresh == *initial_as_of,
+                    "initial_as_of should be set to the first refresh"
+                );
+                self.next_refresh = first_refresh.into();
+            } else {
+                // The first refresh has already happened.
+                let write_frontier = write_frontier.as_option().expect("checked above");
+                self.last_completed_refresh = write_frontier
+                    .round_down_minus_1(&self.refresh_schedule)
+                    .map_or_else(
+                        || {
+                            soft_panic_or_log!(
+                                "rounding down should have returned the first refresh or later"
+                            );
+                            Datum::Null
+                        },
+                        |last_completed_refresh| last_completed_refresh.into(),
+                    );
+                self.next_refresh = write_frontier.clone().into();
+            }
+        }
     }
 }
